requirements.txt
# requirements.txt
playwright==1.42.0
beautifulsoup4==4.12.2
selectolax>=0.3.16
extruct>=0.15.0
w3lib>=2.1.2
python-dotenv>=1.0.1
openai>=1.30.1
PyYAML==6.0.1
lxml>=5.2.1
soupsieve>=2.5
requests==2.31.0
httpx>=0.27.0
certifi>=2024.2.2
aiohttp==3.9.3
python-dateutil==2.8.2
